# api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "G-Zang (归藏) API文档",
    "description": "G-Zang (归藏) 财务管理系统API文档",
    "version": "1.0.0",
    "contact": {
      "name": "G-Zang Team",
      "email": "team@gzang.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发环境"
    }
  ],
  "security": [
    {
      "Bearer Authentication": []
    }
  ],
  "components": {
    "securitySchemes": {
      "Bearer Authentication": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "在请求头中添加 JWT Token：Authorization: Bearer {token}"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "用户ID"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "password": {
            "type": "string",
            "description": "密码"
          },
          "nickname": {
            "type": "string",
            "description": "用户昵称"
          },
          "avatar": {
            "type": "string",
            "description": "用户头像URL"
          },
          "roleId": {
            "type": "integer",
            "format": "int64",
            "description": "角色ID"
          },
          "companyId": {
            "type": "integer",
            "format": "int64",
            "description": "公司ID"
          },
          "status": {
            "type": "integer",
            "description": "状态 (0:禁用, 1:正常)"
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "tags": ["用户管理"],
        "summary": "获取用户列表",
        "description": "分页获取用户列表",
        "parameters": [
          {
            "name": "current",
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "每页大小",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "tags": ["用户管理"],
        "summary": "获取用户信息",
        "description": "根据用户ID获取用户信息",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "用户ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          },
          "404": {
            "description": "用户不存在"
          }
        }
      },
      "put": {
        "tags": ["用户管理"],
        "summary": "更新用户信息",
        "description": "更新指定用户的信息",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "用户ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "delete": {
        "tags": ["用户管理"],
        "summary": "删除用户",
        "description": "删除指定用户",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "用户ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
  }
}